% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_filename.R
\name{la_cache_filename}
\alias{la_cache_filename}
\title{Returns the name of the cache file associated with the given URL. Note that this file 
may not actually exist, this function just provides the mapping from URL to filename}
\usage{
la_cache_filename(url)
}
\arguments{
\item{url}{string: the URL}
}
\value{
string: the file path and name
}
\description{
Returns the name of the cache file associated with the given URL. Note that this file 
may not actually exist, this function just provides the mapping from URL to filename
}
\examples{
la_cache_filename("https://records-ws.nbnatlas.org/index/fields")

}
\seealso{
\code{la_config} for cache settings, particularly the cache directory
}
