% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{la_list}
\alias{la_list}
\title{Species lists}
\usage{
la_list(druid, kvp = TRUE, verbose = la_config()$verbose)
}
\arguments{
\item{druid}{string: data resource UID of the list (i.e. the list identifier)}

\item{kvp}{logical: include key-value pairs? Some lists contain information about the species in the form of key-value pairs}

\item{verbose}{logical: show additional progress information?}
}
\value{
data.frame
}
\description{
Note that this refers to pre-generated lists of species stored on Living Atlas servers. The similarly-named but different function \code{\link{la_specieslist}} provides a different function, namely listing the species matching a query or recorded as present in a search area.
}
\examples{
\dontrun{
 la_config(installation = "au")
 all_lists <- la_lists()
 ## find the "Field Guide apps species profiles" from Museum Victoria
 all_lists[grep("Field Guide", all_lists$listName), ]
 ## download the vertebrates one
 l <- la_list(druid = "dr1146")
}

}
\seealso{
\code{\link{la_specieslist}}, \code{\link{la_lists}}
}
