% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_info.R
\name{la_species_info}
\alias{la_species_info}
\title{Fetch a taxon profile given a scientific name or LSID (GUID)}
\usage{
la_species_info(scientificname, guid, verbose = la_config()$verbose)
}
\arguments{
\item{scientificname}{string: scientific name of the taxon of interest (species, genus, family etc)}

\item{guid}{string: The Life Science Identifier of the taxon of interest}

\item{verbose}{logical: show additional progress information? [default is set by nbn_config()]}
}
\value{
A species profile in the form of a named list, each element of which is generally a data frame. An empty list is returned if no match is found for the supplied name or guid
}
\description{
Fetch a taxon profile given a scientific name or LSID (GUID)
}
\examples{
\dontrun{
 la_config(installation = "uk")
 la_species_info("Diatoma tenuis")
 la_species_info(guid="NHMSYS0000080188")
 la_species_info("Diatoma", verbose = TRUE)
}
}
\seealso{
\code{\link{la_config}}
}
