% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieslist.R
\name{la_specieslist}
\alias{la_specieslist}
\title{Get list of taxa and their occurrence counts}
\usage{
la_specieslist(taxon, wkt, fq)
}
\arguments{
\item{taxon}{string: (optional) query of the form field:value (e.g. "genus:Macropus") or a free text search (e.g. "macropodidae").
For reliable results it is recommended to use a specific field where possible (see \code{la_fields("occurrence_indexed")}
for valid fields). It is also good practice to quote the taxon name if it contains multiple words, for example
\code{taxon="taxon_name:\"Vulpes vulpes\""} (noting, however, that multi-word names are unlikely in the context of a specieslist
search, where one would typically be searching for all species within, say, a genus or family)}

\item{wkt}{string: WKT (well-known text) defining a polygon within which to limit taxon search, e.g. "POLYGON((-3 56,-4 56,-4 57,-3 57,-3 56))"}

\item{fq}{string: character string or vector of strings, specifying filters to be applied to the original query. These are of the form "INDEXEDFIELD:VALUE" e.g. "kingdom:Fungi". See \code{la_fields("occurrence_indexed",as_is=TRUE)} for all the fields that are queryable. NOTE that fq matches are case-sensitive, but sometimes the entries in the fields are not consistent in terms of case (e.g. kingdom names "Fungi" and "Plantae" but "ANIMALIA"). fq matches are ANDed by default (e.g. c("field1:abc","field2:def") will match records that have field1 value "abc" and field2 value "def"). To obtain OR behaviour, use the form c("field1:abc OR field2:def")}
}
\value{
data frame of results, where each row is a taxon, its classification information, and its occurrence count
}
\description{
Retrieve a list of taxa matching a search query, within a spatial search area, or both.
}
\examples{
\dontrun{
la_config(installation = "uk")
x <- la_specieslist(taxon = "genus:Leuctra", wkt = "POLYGON((-3 56,-4 56,-4 57,-3 57,-3 56))")

x <- la_specieslist(wkt = "POLYGON((-3 56,-4 56,-4 57,-3 57,-3 56))", fq = "rank:species")

la_config(installation = "au")
x <- la_specieslist(
    wkt = "POLYGON((152.38 -30.43,152.5 -30.43,152.5 -30.5,152.38 -30.5,152.38 -30.43))",
    fq = "kingdom:Plantae")
## NOTE that this response might include records with empty or NA kingdom, phylum, or
##  class values, as per the note above.
}
}
\references{
\url{http://www.geoapi.org/3.0/javadoc/org/opengis/referencing/doc-files/WKT.html}
}
\seealso{
\code{\link{la_fields}} for occurrence fields that are queryable via the \code{fq} parameter
}
