% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livingatlases.R
\docType{package}
\name{livingatlases}
\alias{livingatlases}
\alias{livingatlases-package}
\title{\pkg{livingatlases}}
\description{
This package enables the R community to access data and tools hosted by the 
Living Atlases project.
}
\references{
\url{https://living-atlases.gbif.org/}
}
